+++ Magic/CHANGELOG.md
@@ -46,6 +46,7 @@
 
 ## Miscellaneous Changes
 
+ - Add "jump" trigger (for players only)
  - Add ignored_by_mobs wand/class/modifier property
  - Add use_target_mage parameter to ChangeContext action
    Useful for actions such as Selector and CheckRequirement that only use the mage context and not the target